backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
qrcode[pil]==7.4.2
websockets==12.0
pillow
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 